Drop IntendedState from resolver
