+++ lib/resolve.go
@@ -1,26 +1,29 @@
 package sous
 
-import (
-	"strings"
-	"sync"
-)
+import "sync"
 
 type (
 	// Resolver is responsible for resolving intended and actual deployment
 	// states.
 	Resolver struct {
-		Deployer      Deployer
-		Registry      Registry
-		IntendedState State
+		Deployer Deployer
+		Registry Registry
 	}
 )
 
-func NewResolver(d Deployer, r Registry, intended State) *Resolver {
+func NewResolver(d Deployer, r Registry) *Resolver {
 	return &Resolver{
-		Deployer:      d,
-		Registry:      r,
-		IntendedState: intended,
+		Deployer: d,
+		Registry: r,
 	}
+}
+
+// Resolve drives the Sous deployment resolution process. It calls out to the
+// appropriate components to compute the intended deployment set, collect the
+// actual set, compute the diffs and then issue the commands to rectify those
+// differences.
+func (r *Resolver) Resolve(intended State) error {
+	return r.ResolveFilteredDeployments(intended, nil)
 }
 
 // Rectify takes a DiffChans and issues the commands to the infrastructure to reconcile the differences
@@ -36,27 +39,11 @@ func Rectify(dcs DiffChans, d Deployer, reg Registry) chan RectificationError {
 	return errs
 }
 
-func (re *ResolveErrors) Error() string {
-	s := []string{"Errors during resolve:"}
-	for _, e := range re.Causes {
-		s = append(s, e.Error())
-	}
-	return strings.Join(s, "\n  ")
-}
-
-// Resolve drives the Sous deployment resolution process. It calls out to the
-// appropriate components to compute the intended deployment set, collect the
-// actual set, compute the diffs and then issue the commands to rectify those
-// differences.
-func (r *Resolver) Resolve() error {
-	return r.ResolveFilteredDeployments(nil)
-}
-
 // ResolveFilteredDeployments is similar to Resolve, but also accepts a
 // predicate to filter those deployments. See Deploments.Filter for details.
-func (r *Resolver) ResolveFilteredDeployments(pr DeploymentPredicate) error {
+func (r *Resolver) ResolveFilteredDeployments(intended State, pr DeploymentPredicate) error {
 	Log.Debug.Print("Loading GDM")
-	gdm, err := r.IntendedState.Deployments()
+	gdm, err := intended.Deployments()
 	gdm = gdm.Filter(pr)
 	if err != nil {
 		return err
@@ -64,7 +51,7 @@ func (r *Resolver) ResolveFilteredDeployments(pr DeploymentPredicate) error {
 
 	Log.Debug.Print("Loaded. Collecting ADC...")
 
-	ads, err := r.Deployer.GetRunningDeployment(r.IntendedState.BaseURLs())
+	ads, err := r.Deployer.GetRunningDeployment(intended.BaseURLs())
 	if err != nil {
 		return err
 	}
@@ -93,15 +80,6 @@ func (r *Resolver) ResolveFilteredDeployments(pr DeploymentPredicate) error {
 	return nil
 }
 
-func (e *MissingImageNamesError) Error() string {
-	causeStrs := make([]string, 0, len(e.Causes)+1)
-	causeStrs = append(causeStrs, "Image names are unknown to Sous for source versions")
-	for _, c := range e.Causes {
-		causeStrs = append(causeStrs, c.Error())
-	}
-	return strings.Join(causeStrs, "  \n")
-}
-
 func guardImageNamesKnown(r Registry, gdm Deployments) error {
 	es := make([]error, 0, len(gdm))
 	for _, d := range gdm {
@@ -115,21 +93,3 @@ func guardImageNamesKnown(r Registry, gdm Deployments) error {
 	}
 	return nil
 }
-
-// ResolveFromDir does everything that Resolve does, plus it adds loading the
-// Sous config from a directory of YAML files. This use case is important for
-// proof-of-concept, but long term we expect to be able to abstract the storage
-// of the Sous state away, so this might be deprecated at some point.
-//func (r *Resolver) ResolveFromDir(dir string) error {
-//	return r.ResolveFromDirFiltered(dir, nil)
-//}
-//
-//// ResolveFromDirFiltered is similar to ResolveFromDir, but additionally filters
-//// the deployments to be resolved based on the predicate.
-//func (r *Resolver) ResolveFromDirFiltered(dir string, pr DeploymentPredicate) error {
-//	config, err := LoadState(dir)
-//	if err != nil {
-//		return err
-//	}
-//	return r.ResolveFilteredDeployments(pr)
-//}
